+++ packages/docs/docs/terminology/entry-point.mdx
@@ -30,6 +30,10 @@ If you call [`npx remotion render`](/docs/cli/render), the entry point and the r
 <li><code>remotion/index.ts</code></li>
 <li><code>remotion/index.js</code></li>
 <li><code>remotion/index.mjs</code></li>
+<li><code>src/remotion/index.tsx</code></li>
+<li><code>src/remotion/index.ts</code></li>
+<li><code>src/remotion/index.js</code></li>
+<li><code>src/remotion/index.mjs</code></li>
 </ul>
 
 ## In the Player
