--- packages/template-next-app/.eslintrc.json
@@ -1,10 +1,0 @@
-{
-  "extends": "next",
-  "plugins": ["@remotion"],
-  "overrides": [
-    {
-      "files": ["remotion/**/*.{ts,tsx}"],
-      "extends": ["plugin:@remotion/recommended"]
-    }
-  ]
-}
