+++ packages/template-next-app/package.json
@@ -6,7 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint .",
+    "lint": "next lint",
     "remotion": "remotion studio",
     "render": "remotion render",
     "deploy": "node deploy.mjs"
@@ -19,7 +19,7 @@
     "@remotion/paths": "workspace:*",
     "@remotion/player": "workspace:*",
     "@remotion/shapes": "workspace:*",
-    "next": "15.1.2",
+    "next": "15.1.6",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "remotion": "workspace:*",
@@ -33,8 +33,8 @@
     "@types/react-dom": "19.0.0",
     "@types/web": "0.0.166",
     "dotenv": "16.0.3",
-    "eslint": "8.57.1",
-    "eslint-config-next": "15.0.4",
+    "eslint": "9.19.0",
+    "@next/eslint-plugin-next": "15.1.6",
     "typescript": "5.5.4",
     "prettier": "3.3.3"
   }
