+++ packages/template-next-app/src/app/api/lambda/progress/route.ts
@@ -3,7 +3,7 @@ import {
   AwsRegion,
   getRenderProgress,
 } from "@remotion/lambda/client";
-import { DISK, RAM, REGION, TIMEOUT } from "../../../../config.mjs";
+import { DISK, RAM, REGION, TIMEOUT } from "../../../../../config.mjs";
 import { executeApi } from "../../../../helpers/api-response";
 import { ProgressRequest, ProgressResponse } from "../../../../types/schema";
