+++ packages/template-next-app/src/app/api/lambda/render/route.ts
@@ -3,8 +3,9 @@ import {
   renderMediaOnLambda,
   speculateFunctionName,
 } from "@remotion/lambda/client";
-import { DISK, RAM, REGION, SITE_NAME, TIMEOUT } from "../../../../config.mjs";
 import { executeApi } from "../../../../helpers/api-response";
+import { TIMEOUT } from "dns";
+import { DISK, RAM, REGION, SITE_NAME } from "../../../../../config.mjs";
 import { RenderRequest } from "../../../../types/schema";
 
 export const POST = executeApi<RenderMediaOnLambdaOutput, typeof RenderRequest>(
