+++ packages/template-next-app/src/app/favicon.ico
