+++ packages/template-next-app/src/components/Error.tsx
