+++ packages/template-next-app/src/components/Input.tsx
