+++ packages/template-next-app/src/components/ProgressBar.tsx
