+++ packages/template-next-app/src/components/RenderControls.tsx
