+++ packages/template-next-app/src/components/Spinner/styles.module.css
