+++ packages/template-next-app/src/components/Tips/Tips.tsx
