+++ packages/template-next-app/src/helpers/api-response.ts
