+++ packages/template-next-app/src/helpers/use-rendering.ts
