+++ packages/template-next-app/src/remotion/MyComp/NextLogo.tsx
