+++ packages/template-next-app/src/remotion/MyComp/Rings.tsx
