+++ packages/template-next-app/src/remotion/MyComp/TextFade.tsx
