+++ packages/template-next-app/src/remotion/Root.tsx
