+++ packages/template-next-app/src/remotion/index.ts
