+++ packages/template-next-app/src/types/constants.ts
