+++ packages/template-next-app/src/types/schema.ts
